Implement socket connection class
