+++ server/browser-management/BrowserSession.ts
@@ -1,9 +1,8 @@
 import { chromium, Page, Browser, CDPSession } from 'playwright';
 import { uuid } from 'uuidv4';
-import { Socket } from 'socket.io';
-import { io } from '../server';
 import * as fs from 'fs';
 import logger from '../logger';
+import {SocketConnection} from "./SocketConnection";
 
 export class BrowserSession {
 
@@ -17,30 +16,14 @@ export class BrowserSession {
 
     private client : CDPSession|null = null;
 
-    private socket : Socket|null = null;
+    private socketConnection : SocketConnection;
 
-    constructor(){
+    public constructor(clientConnection: SocketConnection){
         this.id = uuid();
-        io.on('connection', async (socket: Socket) => {
-            logger.log('info',"Client connected");
-            this.socket = socket;
-            socket.on('disconnect', () => {
-                logger.log('info', "Client disconnected");
-            });
-        });
+        this.socketConnection = clientConnection;
     }
 
-    private emitScreenshot = (payload: any) : void => {
-        if (!this.socket) {
-            logger.log('warn','socket is not connected');
-            return;
-        }
-        const dataWithMimeType = ('data:image/jpeg;base64,').concat(payload);
-        this.socket.emit('screencast', dataWithMimeType);
-        logger.log('debug',`Screenshot emitted`);
-    };
-
-    public async initialize(options: any) : Promise<void> {
+    public initialize = async(options: any) : Promise<void> => {
         // initialize the browser instance
         this.browser = <Browser>(await chromium.launch(
             {
@@ -55,7 +38,7 @@ export class BrowserSession {
         this.client = await this.currentPage.context().newCDPSession(this.currentPage);
     };
 
-    private async startScreencast() : Promise<void> {
+    private startScreencast = async() : Promise<void> => {
         if (!this.client) {
             logger.log('warn','client is not initialized');
             return;
@@ -64,14 +47,14 @@ export class BrowserSession {
         logger.log('info',`BrowserSession with id ${this.id} started with a screencasting.`);
     };
 
-    public async subscribeToScreencast() : Promise<void> {
+    public subscribeToScreencast = async() : Promise<void> => {
         await this.startScreencast();
         if (!this.client) {
             logger.log('warn','client is not initialized');
             return;
         }
         this.client.on('Page.screencastFrame', ({ data: base64, sessionId }) => {
-            this.emitScreenshot(base64);
+            this.socketConnection.emitScreenshot(base64);
             setTimeout(async () => {
                 try {
                     if (!this.client) {
@@ -86,7 +69,7 @@ export class BrowserSession {
         });
     };
 
-    public async stopScreencast() : Promise<void> {
+    public stopScreencast = async() : Promise<void> => {
         if (!this.client) {
             logger.log('warn','client is not initialized');
             return;
@@ -95,13 +78,12 @@ export class BrowserSession {
         logger.log('info',`BrowserSession with id ${this.id} stopped with a screencasting.`);
     };
 
-    public async openPage(url: string) : Promise<void>{
+    public openPage = async(url: string) : Promise<void> =>{
         if (this.currentPage) {
             logger.log('debug',`Page ${url} opened`)
             await this.currentPage.goto(url);
             const image = await this.currentPage.screenshot();
             fs.writeFileSync('screenshot.png', image);
-            // this.emitScreenshot(image);
         } else {
             logger.log('warn','Page is not initialized');
         }
